TutorDexAggregator/requirements.txt
requests>=2.28.0
telethon>=1.31.0
python-dotenv>=1.0.0
json-repair
prometheus-client>=0.20.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
